BuildingProperties.cs
Updated

+++ BuildingProperties.cs
@@ -4,16 +4,16 @@ using UnityEngine;
 
 public class BuildingProperties : MonoBehaviour
 {
-    [SerializeField] public int maxhealth = 3;
+    [SerializeField] public float maxhealth = 3;
 
-    [SerializeField] public int currenthealth;
+    [SerializeField] public float currenthealth;
 
     void Start()
     {
         currenthealth = maxhealth;
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         currenthealth -= damage;
